+++ demo/saxpy/scalar/saxpy_scalar.cpp
@@ -65,7 +65,7 @@ template<typename T> struct axpy_scalar
     std::vector<T> X, Y;
 };
 
-NT2_REGISTER_BENCHMARK_TPL( axpy_scalar, (float) )
+NT2_REGISTER_BENCHMARK_TPL( axpy_scalar, (float)(double) )
 {
   std::size_t size_min = args("size_min", 16);
   std::size_t size_max = args("size_max", 4096);
